requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
chromadb>=0.4.0
pypdf2>=3.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
python-multipart>=0.0.6